refactor: folders of weapons activates
