+++ js/shipUnits/weapons/info/energy/weapons/disassemble1.js
@@ -1,7 +1,7 @@
 
-import { InheritController } from "../../../../generalUtils/inherit.js"
-import { Disassemble1Effect } from "../effects/disassemble1Effect.js"
-import { WeaponExtend } from "../extend/weapon.js"
+import { InheritController } from "../../../../../generalUtils/inherit.js"
+import { Disassemble1Effect } from "../../effects/disassemble1Effect.js"
+import { WeaponExtend } from "../../extend/weapon.js"
 
 export class Disassemble1 {
