+++ js/shipUnits/weapons/info/energy/weapons/fragilizer1.js
@@ -1,7 +1,7 @@
 
-import { InheritController } from "../../../../generalUtils/inherit.js"
-import { Fragilizer1Effect } from "../effects/fragilizer1Effect.js"
-import { WeaponExtend } from "../extend/weapon.js"
+import { InheritController } from "../../../../../generalUtils/inherit.js"
+import { Fragilizer1Effect } from "../../effects/fragilizer1Effect.js"
+import { WeaponExtend } from "../../extend/weapon.js"
 
 export class Fragilizer1 {
