+++ js/shipUnits/weapons/info/energy/weapons/miniWorldLauncher.js
@@ -1,5 +1,5 @@
-import { InheritController } from "../../../../generalUtils/inherit.js"
-import { WeaponExtend } from "../extend/weapon.js"
+import { InheritController } from "../../../../../generalUtils/inherit.js"
+import { WeaponExtend } from "../../extend/weapon.js"
 
 export class MiniWorldLauncher {
