+++ js/shipUnits/weapons/info/energy/weapons/missileBurst1.js
@@ -1,8 +1,8 @@
 
-import { FocusedTopDownBehavior } from "../../../../AI/behavior/focusedTopDownBehavior.js"
-import { InheritController } from "../../../../generalUtils/inherit.js"
-import { HomingWeapon } from "../extend/homingWeapon.js"
-import { WeaponExtend } from "../extend/weapon.js"
+import { FocusedTopDownBehavior } from "../../../../../AI/behavior/focusedTopDownBehavior.js"
+import { InheritController } from "../../../../../generalUtils/inherit.js"
+import { HomingWeapon } from "../../extend/homingWeapon.js"
+import { WeaponExtend } from "../../extend/weapon.js"
 
 export class MissileBurst1 {
