+++ js/shipUnits/weapons/info/energy/weapons/paintbrush1.js
@@ -1,7 +1,7 @@
 
-import { InheritController } from "../../../../generalUtils/inherit.js"
-import { Paintbrush1Effect } from "../effects/paintbrush1Effect.js"
-import { WeaponExtend } from "../extend/weapon.js"
+import { InheritController } from "../../../../../generalUtils/inherit.js"
+import { Paintbrush1Effect } from "../../effects/paintbrush1Effect.js"
+import { WeaponExtend } from "../../extend/weapon.js"
 
 export class Paintbrush1 {
