+++ js/shipUnits/weapons/info/energy/weapons/parasiteInjection1.js
@@ -1,9 +1,9 @@
 
-import { FocusedTopDownBehavior } from "../../../../AI/behavior/focusedTopDownBehavior.js"
-import { InheritController } from "../../../../generalUtils/inherit.js"
-import { ParasiteInjection1Effect } from "../effects/parasiteInjection1Effect.js"
-import { HomingWeapon } from "../extend/homingWeapon.js"
-import { WeaponExtend } from "../extend/weapon.js"
+import { FocusedTopDownBehavior } from "../../../../../AI/behavior/focusedTopDownBehavior.js"
+import { InheritController } from "../../../../../generalUtils/inherit.js"
+import { ParasiteInjection1Effect } from "../../effects/parasiteInjection1Effect.js"
+import { HomingWeapon } from "../../extend/homingWeapon.js"
+import { WeaponExtend } from "../../extend/weapon.js"
 
 export class ParasiteInjection1 {
