+++ js/shipUnits/weapons/info/energy/weapons/piston1.js
@@ -1,6 +1,6 @@
 
-import { InheritController } from "../../../../generalUtils/inherit.js"
-import { WeaponExtend } from "../extend/weapon.js"
+import { InheritController } from "../../../../../generalUtils/inherit.js"
+import { WeaponExtend } from "../../extend/weapon.js"
 
 export class Piston1 {
