+++ js/shipUnits/weapons/info/weaponsInfoController.js
@@ -4,28 +4,28 @@ import { ActivateInfoController } from "../../forAllShipUnits/activateInfoContro
 import { InheritController } from "../../../generalUtils/inherit.js"
 import { FocusedNearBehavior } from "../../../AI/behavior/focusedNearBehavior.js"
 
-import { BlackHoleGenerator1 } from "./weapons/blackHoleGenerator1.js"
-import { MissileBurst1 } from "./weapons/missileBurst1.js"
-import { MineLauncher1 } from "./weapons/mineLauncher1.js"
-import { Piston1 } from "./weapons/piston1.js"
-import { SniperPiston1 } from "./weapons/sniperPiston1.js"
-import { Scrapper1 } from "./weapons/scrapper1.js"
-import { FlameThrower1 } from "./weapons/flameThrower1.js"
-import { Fragilizer1 } from "./weapons/fragilizer1.js"
-import { Disassemble1 } from "./weapons/disassemble1.js"
-import { ElectrifiedMissile1 } from "./weapons/electrifiedMissile1.js"
-import { Paintbrush1 } from "./weapons/paintbrush1.js"
-import { Missilepiston } from "./weapons/missilepiston.js"
-import { Shotgun1 } from "./weapons/shotgun1.js"
-import { MissileCluster1 } from "./weapons/missileCluster1.js"
-import { MiniWorldLauncher } from "./weapons/miniWorldLauncher.js"
-import { SmallBulletCluster1 } from "./weapons/smallBulletCluster1.js"
-import { Diffusion1 } from "./weapons/diffusion1.js"
-import { Bubbler1 } from "./weapons/bubbler1.js"
-import { ElectrifiedBomb1 } from "./weapons/electrifiedBomb1.js"
-import { ToyMachinegun } from "./weapons/toyMachinegun.js"
-import { ParasiteInjection1 } from "./weapons/parasiteInjection1.js"
-import { DeathHandLauncher } from "./weapons/deathHandLauncher.js"
+import { BlackHoleGenerator1 } from "./energy/weapons/blackHoleGenerator1.js"
+import { MissileBurst1 } from "./energy/weapons/missileBurst1.js"
+import { MineLauncher1 } from "./energy/weapons/mineLauncher1.js"
+import { Piston1 } from "./energy/weapons/piston1.js"
+import { SniperPiston1 } from "./energy/weapons/sniperPiston1.js"
+import { Scrapper1 } from "./energy/weapons/scrapper1.js"
+import { FlameThrower1 } from "./energy/weapons/flameThrower1.js"
+import { Fragilizer1 } from "./energy/weapons/fragilizer1.js"
+import { Disassemble1 } from "./energy/weapons/disassemble1.js"
+import { ElectrifiedMissile1 } from "./energy/weapons/electrifiedMissile1.js"
+import { Paintbrush1 } from "./energy/weapons/paintbrush1.js"
+import { Missilepiston } from "./energy/weapons/missilepiston.js"
+import { Shotgun1 } from "./energy/weapons/shotgun1.js"
+import { MissileCluster1 } from "./energy/weapons/missileCluster1.js"
+import { MiniWorldLauncher } from "./energy/weapons/miniWorldLauncher.js"
+import { SmallBulletCluster1 } from "./energy/weapons/smallBulletCluster1.js"
+import { Diffusion1 } from "./energy/weapons/diffusion1.js"
+import { Bubbler1 } from "./energy/weapons/bubbler1.js"
+import { ElectrifiedBomb1 } from "./energy/weapons/electrifiedBomb1.js"
+import { ToyMachinegun } from "./energy/weapons/toyMachinegun.js"
+import { ParasiteInjection1 } from "./energy/weapons/parasiteInjection1.js"
+import { DeathHandLauncher } from "./energy/weapons/deathHandLauncher.js"
